Change return type for arp_cache_cmp()

+++ lib/src/arp.h
@@ -103,7 +103,7 @@ struct arp_entry {
 };
 
 
-extern int64_t __attribute__((nonnull))
+extern int32_t __attribute__((nonnull))
 arp_cache_cmp(const struct arp_entry * const a,
               const struct arp_entry * const b);
